Initial commit to the project
This is a very early release of the Markdown-based blog module, one that
does just hte basics, which is converting a series of, minimalist,
Markdown files with Yaml front-matter into an array of BlogArticle
entities. These entities can be used to provide a list of blog articles,
or to render a given article.