Extract the functionality to filter published and upcoming items
After looking over and over at how the class was previously written, I
didn't like (rightly or wrongly) how the functionality for retrieving
upcoming and published items was contained directly within
ItemListerFilesystem. Something about it just seemed wrong to me. So,
this change removes it, showing examples of it in the respective tests.
This feels better to me, and allows for any number of different types of
filtering to happen later, if and when required.