Refine the BlogArticle and ItemListerFilesystem tests
After reviewing the tests I saw that they contained tests that were
either unnecessary or duplicates, effectively, and could also be done in
much more efficient ways. Given that, the unnecessary and duplicate ones
were removed, and a dataProvider was used to simplify running multiple
tests in a more efficient way.