Refactor for simpler hydration, filtering, and validation
This is a larger commit, that I should have broken up into small
commits, or done progressively, instead of at one time. The change
refactors BlogArticle to be hydratable using the laminas hydrators, and
integrates the use of an inputfilter to begin to filter and validate the
input used to hydrate BlogArticle object. My intent here is to ensure
that the input data is never blindly trusted, but sanitised
appropriately.