Write an error log message when a blog item cannot be hydrated
For the longest time, I've had issues knowing if errors were caused
because of not being able to hydrate a blog item from a Markdown file,
and if so, why. This change starts to get rid of that issue, by writing
an error log message if an item cannot be hydrated, along with the
reasons why.