Add an iterator to retrieve related posts for an article
I don't know if this is the right or best way to approach this
situation, but it does work. It applies an implementation of a
FilterIterator to determine, simplistically, which articles are related
to the current one. This is only based on matching the tags and
categories between two articles.