Refactor FilterPostBy* iterators to handle null values
I'm completely lost as to why, but for some reason I can replicate the
error in a scratch file and it happens when the iterators are used, but
not during a test. That issue is that if the array passed to array_map
contains a null value, then that value will break the call to
strtolower, the callback passed to array_map. So, because I know that it
happens, I'm implementing this fix and will figure out how to get the
tests to reflect that in time.